+++ packages/common/src/utils/conditionalFormatting.ts
@@ -14,8 +14,13 @@ import {
 import {
     Field,
     FilterableItem,
+    Format,
+    isField,
     isFilterableItem,
+    isMetric,
+    isTableCalculation,
     TableCalculation,
+    TableCalculationFormatType,
 } from '../types/field';
 import { FieldTarget } from '../types/filter';
 import assertUnreachable from './assertUnreachable';
@@ -53,32 +58,59 @@ export const createConditionalFormattingConfigWithColorRange = (
     },
 });
 
+export const hasPercentageFormat = (
+    field: Field | TableCalculation | undefined,
+) => {
+    if (!field) return false;
+
+    return (
+        (isField(field) && field?.format === Format.PERCENT) ||
+        (isTableCalculation(field) &&
+            field.format?.type === TableCalculationFormatType.PERCENT)
+    );
+};
+
+const convertFormattedValue = (
+    value: unknown,
+    field: Field | TableCalculation | undefined,
+) => {
+    if (!field) return value;
+
+    if (hasPercentageFormat(field)) {
+        return typeof value === 'number' ? value * 100 : value;
+    }
+
+    return value;
+};
+
 export const hasMatchingConditionalRules = (
+    field: Field | TableCalculation,
     value: unknown,
     config: ConditionalFormattingConfig | undefined,
 ) => {
     if (!config) return false;
 
     const parsedValue = typeof value === 'string' ? Number(value) : value;
+    const convertedValue = convertFormattedValue(parsedValue, field);
 
     if (isConditionalFormattingConfigWithSingleColor(config)) {
         return config.rules.every((rule) => {
             switch (rule.operator) {
                 case ConditionalOperator.NULL:
-                    return parsedValue === null;
+                    return convertedValue === null;
                 case ConditionalOperator.NOT_NULL:
-                    return parsedValue !== null;
+                    return convertedValue !== null;
                 case ConditionalOperator.EQUALS:
-                    return rule.values.some((v) => parsedValue === v);
+                    return rule.values.some((v) => convertedValue === v);
                 case ConditionalOperator.NOT_EQUALS:
-                    return rule.values.some((v) => parsedValue !== v);
+                    return rule.values.some((v) => convertedValue !== v);
                 case ConditionalOperator.LESS_THAN:
-                    return typeof parsedValue === 'number'
-                        ? rule.values.some((v) => parsedValue < v)
+                    return typeof convertedValue === 'number'
+                        ? rule.values.some((v) => convertedValue < v)
                         : false;
                 case ConditionalOperator.GREATER_THAN:
-                    return typeof parsedValue === 'number'
-                        ? rule.values.some((v) => parsedValue > v)
+                    return typeof convertedValue === 'number'
+                        ? rule.values.some((v) => convertedValue > v)
                         : false;
                 case ConditionalOperator.STARTS_WITH:
                 case ConditionalOperator.ENDS_WITH:
@@ -102,9 +134,12 @@ export const hasMatchingConditionalRules = (
     }
 
     if (isConditionalFormattingConfigWithColorRange(config)) {
-        if (typeof parsedValue !== 'number') return false;
+        if (typeof convertedValue !== 'number') return false;
 
-        return parsedValue >= config.rule.min && parsedValue <= config.rule.max;
+        return (
+            convertedValue >= config.rule.min &&
+            convertedValue <= config.rule.max
+        );
     }
 
     return assertUnreachable(config, 'Unknown conditional formatting config');
@@ -124,7 +159,7 @@ export const getConditionalFormattingConfig = (
 
     return fieldConfigs
         .reverse()
-        .find((c) => hasMatchingConditionalRules(value, c));
+        .find((config) => hasMatchingConditionalRules(field, value, config));
 };
 
 export const getConditionalFormattingDescription = (
@@ -166,6 +201,7 @@ export const getConditionalFormattingDescription = (
 };
 
 export const getConditionalFormattingColor = (
+    field: Field | TableCalculation | undefined,
     value: unknown,
     conditionalFormattingConfig: ConditionalFormattingConfig | undefined,
     getColorFromRange: (
@@ -191,9 +227,11 @@ export const getConditionalFormattingColor = (
         isConditionalFormattingConfigWithColorRange(conditionalFormattingConfig)
     ) {
         const numericValue = typeof value === 'string' ? Number(value) : value;
-        if (typeof numericValue !== 'number') return undefined;
+        const convertedValue = convertFormattedValue(numericValue, field);
 
-        return getColorFromRange(numericValue, conditionalFormattingConfig);
+        if (typeof convertedValue !== 'number') return undefined;
+
+        return getColorFromRange(convertedValue, conditionalFormattingConfig);
     }
 
     if (